feature(ui-design): bootsrap added and webpack bundler removed bootstrap added to improve ui and webpack and gulp deactivated. Fixes #1003
