+++ gulpfile.js
@@ -44,8 +44,8 @@ gulp.task('scripts', function() {
     .pipe(jshint.reporter('default'))
     .pipe(concat('main.js'))
     .pipe(gulp.dest('dist/assets/js'))
-    .pipe(rename({suffix: '.min'}))
-    .pipe(uglify())
+    // .pipe(rename({suffix: '.min'}))
+    // .pipe(uglify())
     .pipe(gulp.dest('dist/assets/js'))
     .pipe(notify({ message: 'Scripts task complete' }));
 });
@@ -140,7 +140,7 @@ var path = {
  return gulp.src(path.ALL) // gulp looks for all source files under specified path
  .pipe(sourcemaps.init()) // creates a source map which would be very helpful for debugging by maintaining the actual source code structure
  .pipe(stream(webpackConfig)) // blend in the webpack config into the source files
- .pipe(uglify())// minifies the code for better compression
+ // .pipe(uglify())// minifies the code for better compression
  .pipe(sourcemaps.write())
  .pipe(gulp.dest(path.DEST_BUILD));
  });
