+++ root.js
@@ -1,9 +1,9 @@
-angular.module("root", [])
-	.controller("index", ["$scope", function($scope) {
-		$scope.products = [
-			{id: 1, name: "Hockey puck"},
-			{id: 2, name: "Golf club"},
-			{id: 3, name: "Baseball bat"},
-			{id: 4, name: "Lacrosse stick"}
-		];
-}]);
+// angular.module("root", [])
+// 	.controller("index", ["$scope", function($scope) {
+// 		$scope.products = [
+// 			{id: 1, name: "Hockey puck"},
+// 			{id: 2, name: "Golf club"},
+// 			{id: 3, name: "Baseball bat"},
+// 			{id: 4, name: "Lacrosse stick"}
+// 		];
+// }]);
